Refactor database schema and transaction handling for SQLite compatibility
- Updated financial schema to use SQLite types and methods, replacing PostgreSQL specific types and functions.
- Refactored transaction handling to use a synchronous approach with manual BEGIN/COMMIT/ROLLBACK for SQLite.
- Removed deprecated API endpoints and consolidated asset-related functionalities into a single API surface using Elysia.
- Adjusted integration tests to work with SQLite and ensure transaction integrity.
- Updated SvelteKit configuration to use Node adapter instead of Cloudflare.
- Removed unnecessary files and cleaned up imports across the codebase.

+++ src/lib/server/repositories/accounts.ts
@@ -36,15 +36,15 @@ export async function listAccountsWithUsage(userId: string, db: DbClient = read(
 			isActive: financialAccounts.isActive,
 			createdAt: financialAccounts.createdAt,
 			assetCount: sql<number>`(
-				select count(*)::int from ${assets}
+				select count(*) from ${assets}
 				where ${assets.accountId} = ${financialAccounts.id}
 			)`,
 			liabilityCount: sql<number>`(
-				select count(*)::int from ${liabilities}
+				select count(*) from ${liabilities}
 				where ${liabilities.accountId} = ${financialAccounts.id}
 			)`,
 			transactionCount: sql<number>`(
-				select count(*)::int from ${transactions}
+				select count(*) from ${transactions}
 				where ${transactions.accountId} = ${financialAccounts.id}
 			)`
 		})
